Added lab2b to the repo

+++ lab2a.cpp
@@ -7,13 +7,13 @@ int main()
    cout << "Here are some approximations of PI:" << endl;
 
    // Archimedes 225 BC
-   cout << "22 / 7 = " << (float)22 / 7 << endl;
+   cout << "22 / 7 = " << (double)22 / 7 << endl;
 
    // Zu Chongzhi 480 AD
-   cout << "355 / 113 = " << (float)355 / 113 << endl;
+   cout << "355 / 113 = " << (double)355 / 113 << endl;
 
    // Indiana law 1897 AD
-   cout << "16 / 5 = " << (float)16 / 5 << endl;
+   cout << "16 / 5 = " << (double)16 / 5 << endl;
 
    return 0;
 }
